+++ front/pages/api/w/[wId]/groups.ts
@@ -8,18 +8,11 @@ import type { Authenticator } from "@app/lib/auth";
 import { GroupResource } from "@app/lib/resources/group_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { GroupKind, GroupType, WithAPIErrorResponse } from "@app/types";
+import { GroupKindCodec } from "@app/types";
 
 export type GetGroupsResponseBody = {
   groups: GroupType[];
 };
-
-export const GroupKindCodec = t.keyof({
-  global: null,
-  regular: null,
-  agent_editors: null,
-  system: null,
-  provisioned: null,
-});
 
 const GetGroupsQuerySchema = t.partial({
   kind: t.union([GroupKindCodec, t.array(GroupKindCodec)]),
